Refactor completion storage and add some handling for module arguments
"Leaf modules" are modules that can't take a body, such as primitive
shapes. They are completed with snippets containing trailing semicolons
and a configurable set of arguments. "Group modules" are modules that do
take a body. Their snippets contain braces for the body, within which
the cursor ends up afterward.

Right now, we just bake in hardcoded arguments for some of the built-in
modules. Extracting arguments from user-defined functions/modules is
future work.