Parse module/function declarations and define builtins in-language
Slight regression in that we now can't easily tell whether a module is a
leaf or not, but we definitely want to move away from the big static
list.

It might work to check for the presence of `children` in the body, as
hinted at by the current contents of `builtins.scad`, but handling the
nested case could blow up in complexity. Or we could use an internal
naming convention for builtins and give up on user-defined code. Or we
could look for special marker comments.