Handle include/use statements
The strategy turns out to be pretty natural: when we hit an
`include`/`use` node while looking for visible items, do a recursive
lookup on the referenced file (with the position set to the very start
of the file, which yields all top-level constructs) and dump the results
into the current list (filtered to exclude variables if it's a `use`).

Not handling library search paths just yet, because I don't feel like
dealing with platform-specific default paths right now.

I believe the use of `Rc<RefCell<ParsedCode>>` is necessary because we
may now need to update the map of code (to read a file from disk) while
we have some of it borrowed, though I'm not totally sure there isn't
some tricky way to get around that.